add: sending and delivery pages
